Near completed functionality for the Mood Tracker Calendar!
Completely redid the Mood Tracker Calendar and the "Did you miss anything?" bar.

Users can now select a Mood upon entering the Dashboard, and the Mood will show up on the calendar for the current day.

Users can also click on any other day of the current month to log their Mood for that day. There's an error that all other days are given the same Mood even if different Moods are logged for different days, but I will fix that very soon. Just wanted this commit to go up now.