Updated the Mood Tracker Calendar
There's still a bug in the calendar: If a user were to log moods for multiple days (aside from the current day), all of the emojis would change to the most recently logged one. For whatever reason, the JQuery click() is saving the IDs of previously selected/clicked days. I'm not sure how to fix it, but I gave back the functionality to at least change log a Mood for one other day other than the current day for now.